Remove ArgumentNullException If No ContentId
This removes the intentional throwing of ArgumentNullException, when
the Attachment type is inline, but there is no ContentId.

In my migration of previous emails stored in 'Outlook Express' and
'Windows Live Mail', I found this condition was often hit, thus causing
a failure of the message to be converted to MSG file format.

In my review of https://www.rfc-editor.org/rfc/rfc3801.txt, there were
numerous examples of message/disposition-notification sections without
a ContentId.  In my own testing, setting attachments without a
ContentId, with out throwing this exception, the email went on later
to be seemingly parsed just fine.